Release 7.0.0 (#151)
* Update CleverTap package (#143)

* Update import package script

* Update CleverTap package

* Use EDM4U (#144)

* Import EDM4U

* Add iOS dependencies

* Remove unused android dependencies

* Remove xcframeworks

* Update project version

* Update build

* Update package (#145)

* Update iOS CleverTap initialization (#146)

* Update iOS CleverTap initialization

* Update package

* Make LP Android plugin a library module (#147)

Remove LP Android aar build and instead add it as a source module to the
unityLibrary Android module. Make it depend on the clevertap library module
so its methods can be used in LP Unity plugin.

* Set the CT instance created from native LP Android to CT Unity (#148)

* Set the CT instance created from native LP Android to CT Unity

* Init CleverTapUnityPlugin on UnityBridge.start

* Add gcm_defaultSenderId to the generated google services xml file

This is needed for CT push notifications to work if the client has not
integrated Firebase Unity package in their project

* Remove application definition from the manifest

* Update unitypackage

* Add documentation
Add Android instructions, Android migration to v7 and
CleverTap integration docs

* Set LP Android version to 7.6.3

* Update packages (#149)

* Update leanplum package (#150)

* Update Leanplum package

* Update iOS SDK version

* Update version and package

* Update docs

---------

+++ Leanplum-Unity-SDK/Assets/Editor/PackageImporter.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.IO;
-using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 using UnityEditor;
@@ -8,42 +7,33 @@ using UnityEngine;
 
 namespace Leanplum.Private
 {
+    [InitializeOnLoad]
     public static class PackageImporter
     {
-        private const string CLEVERTAP_UNITY_VERSION = "3.0.0";
-
-        private static readonly string CLEVERTAP_UNITY_PACKAGE = "CleverTapSDK.unitypackage";
+        private const string CLEVERTAP_UNITY_VERSION = "5.1.0";
+        private static readonly string CLEVERTAP_UNITY_PACKAGE_NAME = "CleverTapSDK";
+        private static readonly string CLEVERTAP_UNITY_PACKAGE = $"{CLEVERTAP_UNITY_PACKAGE_NAME}.unitypackage";
         private static readonly string CLEVERTAP_UNITY_PACKAGE_URL =
             $"https://github.com/CleverTap/clevertap-unity-sdk/raw/{CLEVERTAP_UNITY_VERSION}/{CLEVERTAP_UNITY_PACKAGE}";
+        private static readonly string ASSETS_CLEVERTAP_UNITY = "Assets/CleverTap/";
 
-        private static readonly string ASSETS_CLEVERTAP_UNITY = "Assets/CleverTapUnity/";
-
-        private static readonly string[] FRAMEWORKS_PATHS_TO_REMOVE = new string[]
-        {
-            Path.Combine(Environment.CurrentDirectory, $"{ASSETS_CLEVERTAP_UNITY}/iOS/SDWebImage.framework"),
-            Path.Combine(Environment.CurrentDirectory, $"{ASSETS_CLEVERTAP_UNITY}/iOS/CleverTapSDK.framework")
-        };
-
-        private static string CleverTapAndroidSDKArchivePath
-        {
-            get
-            {
-                string androidAssets = Path.Combine(Environment.CurrentDirectory, $"{ASSETS_CLEVERTAP_UNITY}/Android/");
-                string file = Directory.GetFiles(androidAssets, "clevertap-android-sdk-*.aar").FirstOrDefault();
-                if (!string.IsNullOrEmpty(file))
-                {
-                    return Path.Combine(Environment.CurrentDirectory, $"{ASSETS_CLEVERTAP_UNITY}/Android/", file);
-                }
-
-                return file;
-            }
-        }
+        private const string EDM4U_VERSION = "1.2.185";
+        private static readonly string EDM4U_UNITY_PACKAGE_NAME = $"external-dependency-manager-{EDM4U_VERSION}";
+        private static readonly string EDM4U_UNITY_PACKAGE = $"{EDM4U_UNITY_PACKAGE_NAME}.unitypackage";
+        private static readonly string EDM4U_UNITY_PACKAGE_URL =
+            $"https://github.com/googlesamples/unity-jar-resolver/raw/v{EDM4U_VERSION}/external-dependency-manager-{EDM4U_VERSION}.unitypackage";
+        private static readonly string ASSETS_EDM4U = "Assets/ExternalDependencyManager/";
 
         private static readonly HttpClient httpClient;
 
         static PackageImporter()
         {
             httpClient = new HttpClient();
+
+            AssetDatabase.importPackageStarted += OnImportPackageStarted;
+            AssetDatabase.importPackageCancelled += OnImportPackageCancelled;
+            AssetDatabase.importPackageCompleted += OnImportPackageCompleted;
+            AssetDatabase.importPackageFailed += OnImportPackageFailed;
         }
 
         [MenuItem(MenuConstants.LEANPLUM_TOOLS_MENU + "Import CleverTap Package " + CLEVERTAP_UNITY_VERSION)]
@@ -58,16 +48,38 @@ namespace Leanplum.Private
             if (!successful)
                 return;
 
+            // Delete CleverTap package folder.
+            // This ensures files deleted in the new package version are not left in project.
+            if (Directory.Exists(ASSETS_CLEVERTAP_UNITY))
+            {
+                Directory.Delete(ASSETS_CLEVERTAP_UNITY, true);
+            }
+
             // Import the package
-            Task<string> importPackageTask = ImportPackageAsync(packagePath);
-            string result = await importPackageTask;
-            if (string.IsNullOrEmpty(result))
+            AssetDatabase.ImportPackage(packagePath, false);
+        }
+
+        [MenuItem(MenuConstants.LEANPLUM_TOOLS_MENU + "Import EDM4U " + EDM4U_VERSION)]
+        public static async void ImportEDM4U()
+        {
+            Debug.Log($"Import EDM4U Package {EDM4U_VERSION} Started");
+            // Path to save the downloaded package
+            string packagePath = Path.Combine(Environment.CurrentDirectory, EDM4U_UNITY_PACKAGE);
+            // Download the package
+            Task<bool> downloadPackageTask = DownloadFileAsync(EDM4U_UNITY_PACKAGE_URL, packagePath);
+            bool successful = await downloadPackageTask;
+            if (!successful)
                 return;
 
-            // Delete files from the CleverTap package that the Leanplum SDK will provide by itself
-            // Delete the downloaded package file
-            RemoveUnnecessaryFiles(packagePath);
-            Debug.Log($"Import CleverTap Package {CLEVERTAP_UNITY_VERSION} Completed");
+            // Delete CleverTap package folder.
+            // This ensures files deleted in the new package version are not left in project.
+            if (Directory.Exists(ASSETS_EDM4U))
+            {
+                Directory.Delete(ASSETS_EDM4U, true);
+            }
+
+            // Import the package
+            AssetDatabase.ImportPackage(packagePath, false);
         }
 
         private static async Task<bool> DownloadFileAsync(string url, string savePath)
@@ -96,72 +108,45 @@ namespace Leanplum.Private
             }
         }
 
-        private static void RemoveUnnecessaryFiles(string packagePath)
+        private static void OnImportPackageStarted(string packageName)
         {
-            foreach (string path in FRAMEWORKS_PATHS_TO_REMOVE)
-            {
-                if (Directory.Exists(path))
-                {
-                    Directory.Delete(path, true);
-                }
-            }
-
-            string archivePath = CleverTapAndroidSDKArchivePath;
-            if (!string.IsNullOrEmpty(archivePath) && File.Exists(archivePath))
-            {
-                File.Delete(archivePath);
-            }
-
-            if (File.Exists(packagePath))
-            {
-                File.Delete(packagePath);
-            }
+            Debug.Log($"Import Package Started: {packageName}");
         }
 
-        public static async Task<string> ImportPackageAsync(string packagePath)
+        private static void OnImportPackageCancelled(string packageName)
         {
-            TaskCompletionSource<string> taskCompletionSource = new TaskCompletionSource<string>();
-            try
-            {
-                AssetDatabase.importPackageStarted += OnImportPackageStarted;
-                AssetDatabase.importPackageCancelled += OnImportPackageCancelled;
-                AssetDatabase.importPackageCompleted += OnImportPackageCompleted;
-                AssetDatabase.importPackageFailed += OnImportPackageFailed;
-                AssetDatabase.ImportPackage(packagePath, false);
+            Debug.LogError($"Import Package Cancelled: {packageName}");
+            CleanUp(packageName);
+        }
 
-                return await taskCompletionSource.Task;
-            }
-            catch (Exception exception)
-            {
-                Debug.LogException(exception);
-                return null;
-            }
-            finally
-            {
-                AssetDatabase.importPackageCancelled -= OnImportPackageCancelled;
-                AssetDatabase.importPackageCompleted -= OnImportPackageCompleted;
-                AssetDatabase.importPackageFailed -= OnImportPackageFailed;
-            }
+        private static void OnImportPackageCompleted(string packageName)
+        {
+            Debug.Log($"Import Package Completed: {packageName}");
+            CleanUp(packageName);
+        }
 
-            void OnImportPackageStarted(string packageName)
-            {
-                Debug.Log($"Import Package Started: {packageName}");
-            }
+        private static void OnImportPackageFailed(string packageName, string errorMessage)
+        {
+            Debug.LogError($"Import Package Failed: {packageName}. Error: {errorMessage}");
+            CleanUp(packageName);
+        }
 
-            void OnImportPackageCancelled(string packageName)
+        private static void CleanUp(string packageName)
+        {
+            // Delete the downloaded package file
+            string packagePath = string.Empty;
+            if (packageName == CLEVERTAP_UNITY_PACKAGE_NAME)
             {
-                taskCompletionSource.SetCanceled();
+                packagePath = Path.Combine(Environment.CurrentDirectory, CLEVERTAP_UNITY_PACKAGE);
+            }
+            else if (packageName == EDM4U_UNITY_PACKAGE_NAME)
+            {
+                packagePath = Path.Combine(Environment.CurrentDirectory, EDM4U_UNITY_PACKAGE);
             }
 
-            void OnImportPackageCompleted(string packageName)
+            if (!string.IsNullOrEmpty(packagePath) && File.Exists(packagePath))
             {
-                Debug.Log($"Import Package Completed: {packageName}");
-                taskCompletionSource.SetResult(packageName);
-            }
-
-            void OnImportPackageFailed(string packageName, string errorMessage)
-            {
-                taskCompletionSource.SetException(new Exception(errorMessage));
+                File.Delete(packagePath);
             }
         }
     }
